requirements.txt
fastapi
uvicorn[standard]
openai
pydantic
pydantic-settings
pydantic[email]
aiohttp
redis
python-dotenv
cloudinary
python-multipart
requests
SQLAlchemy>=2.0
asyncpg
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
filetype==1.2.0
Pillow==10.0.0
aiosqlite>=0.17.0
greenlet>=2.0.0